@@ -1,24 +1,32 @@
-task wrapper(type: Wrapper) {
-    gradleVersion = '1.6'
-}
-
 buildscript {
+    ext.kotlin_version = '1.1.3-eap-85'
     repositories {
-        mavenCentral()
+        google()
+        jcenter()
+        maven { url "https://dl.bintray.com/kotlin/kotlin-eap-1.1"}
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:+'
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+
+
+        classpath project.hasProperty('global_gradleAndroidPlugin') ? global_gradleAndroidPlugin : 'com.android.tools.build:gradle:3.2.0'
+        classpath 'com.google.gms:google-services:3.0.0'
+        classpath 'com.google.appengine:gradle-appengine-plugin:1.9.28'
     }
 }
-apply plugin: 'com.android.application'
 
-repositories {
-    mavenCentral()
-}
+apply plugin: 'com.android.application'
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-android-extensions'
 
 dependencies {
+    compile 'com.makeramen:roundedimageview:+'
     compile 'com.github.chrisbanes.photoview:library:+'
-    compile project(':ion:ion')
+    compile 'com.flaviofaria:kenburnsview:1.0.6'
+    compile project (':ion:ion-kotlin')
+
+    compile 'com.google.android.gms:play-services-base:11.0.0'
+    compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
 }
 
 android {
@@ -27,15 +35,16 @@ android {
             manifest.srcFile 'AndroidManifest.xml'
             res.srcDirs = ['res/']
             java.srcDirs = ['src/']
+            assets.srcDirs = ['assets/']
         }
     }
 
     defaultConfig {
-        targetSdkVersion 21
-        minSdkVersion 9
+        targetSdkVersion 23
+        minSdkVersion 14
     }
 
-    compileSdkVersion 21
-    buildToolsVersion "21.1.0"
+    compileSdkVersion project.hasProperty('global_compileSdkVersion') ? global_compileSdkVersion : 25
+    buildToolsVersion project.hasProperty('global_buildToolsVersion') ? global_buildToolsVersion : '25.0.2'
 }
 
