@@ -14,6 +14,7 @@ import com.koushikdutta.async.util.FileCache;
 import com.koushikdutta.async.util.StreamUtility;
 import com.koushikdutta.ion.bitmap.BitmapInfo;
 import com.koushikdutta.ion.gif.GifDecoder;
+import com.koushikdutta.ion.gif.GifFrame;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -23,19 +24,24 @@ import java.nio.ByteBuffer;
  * Created by koush on 1/5/14.
  */
 @TargetApi(Build.VERSION_CODES.GINGERBREAD_MR1)
-public class LoadDeepZoom extends LoadBitmapEmitter implements FutureCallback<File> {
+public class LoadDeepZoom extends LoadBitmapEmitter implements FutureCallback<Response<File>> {
     FileCache fileCache;
-    public LoadDeepZoom(Ion ion, String urlKey, boolean animateGif, IonRequestBuilder.EmitterTransform<File> emitterTransform, FileCache fileCache) {
-        super(ion, urlKey, true, animateGif, emitterTransform);
+    public LoadDeepZoom(Ion ion, String urlKey, boolean animateGif, FileCache fileCache) {
+        super(ion, urlKey, true, animateGif);
         this.fileCache = fileCache;
     }
 
     @Override
-    public void onCompleted(Exception e, final File tempFile) {
+    public void onCompleted(Exception e, final Response<File> response) {
+        if (e == null)
+            e = response.getException();
+
         if (e != null) {
             report(e, null);
             return;
         }
+
+        final File tempFile = response.getResult();
 
         if (ion.bitmapsPending.tag(key) != this) {
 //            Log.d("IonBitmapLoader", "Bitmap load cancelled (no longer needed)");
@@ -62,9 +68,10 @@ public class LoadDeepZoom extends LoadBitmapEmitter implements FutureCallback<Fi
                     if (animateGif && TextUtils.equals("image/gif", options.outMimeType)) {
                         fin = fileCache.get(key);
 
-                        BitmapInfo info = new BitmapInfo(key, options.outMimeType, null, size);
-                        info.gifDecoder = new GifDecoder(ByteBuffer.wrap(StreamUtility.readToEndAsArray(fin)));
-                        info.gifDecoder.nextFrame();
+                        GifDecoder gifDecoder = new GifDecoder(ByteBuffer.wrap(StreamUtility.readToEndAsArray(fin)));
+                        GifFrame frame = gifDecoder.nextFrame();
+                        BitmapInfo info = new BitmapInfo(key, options.outMimeType, frame.image, size);
+                        info.gifDecoder = gifDecoder;
                         report(null, info);
                         return;
                     }
@@ -77,7 +84,7 @@ public class LoadDeepZoom extends LoadBitmapEmitter implements FutureCallback<Fi
                     BitmapInfo info = new BitmapInfo(key, options.outMimeType, bitmap, size);
                     info.decoder = decoder;
                     info.decoderFile = file;
-                    info.loadedFrom = Loader.LoaderEmitter.LOADED_FROM_NETWORK;
+                    info.servedFrom = response.getServedFrom();
                     report(null, info);
                 } catch (Exception e) {
                     report(e, null);
