@@ -9,6 +9,7 @@ import com.koushikdutta.async.future.FutureCallback;
 import com.koushikdutta.async.http.AsyncHttpRequest;
 import com.koushikdutta.async.stream.InputStreamDataEmitter;
 import com.koushikdutta.ion.Ion;
+import com.koushikdutta.ion.ResponseServedFrom;
 import com.koushikdutta.ion.bitmap.BitmapInfo;
 
 import java.io.InputStream;
@@ -19,7 +20,7 @@ import java.io.InputStream;
 public class AssetLoader extends StreamLoader {
     @Override
     public Future<BitmapInfo> loadBitmap(final Context context, final Ion ion, final String key, final String uri, final int resizeWidth, final int resizeHeight, final boolean animateGif) {
-        if (!uri.startsWith("file://android_asset/"))
+        if (!uri.startsWith("file:///android_asset/"))
             return null;
 
         return super.loadBitmap(context, ion, key, uri, resizeWidth, resizeHeight, animateGif);
@@ -27,12 +28,12 @@ public class AssetLoader extends StreamLoader {
 
     @Override
     protected InputStream getInputStream(Context context, String uri) throws Exception {
-        return context.getAssets().open(Uri.parse(uri).getPath().substring(1));
+        return context.getAssets().open(Uri.parse(uri).getPath().replaceFirst("^/android_asset/", ""));
     }
 
     @Override
     public Future<DataEmitter> load(final Ion ion, final AsyncHttpRequest request, final FutureCallback<LoaderEmitter> callback) {
-        if (!request.getUri().getScheme().startsWith("file://android_asset/"))
+        if (request.getUri().getScheme() == null || !request.getUri().toString().startsWith("file:///android_asset/"))
             return null;
 
         final InputStreamDataEmitterFuture ret = new InputStreamDataEmitterFuture();
@@ -40,13 +41,13 @@ public class AssetLoader extends StreamLoader {
             @Override
             public void run() {
                 try {
-                    InputStream stream = ion.getContext().getContentResolver().openInputStream(Uri.parse(request.getUri().toString()));
+                    InputStream stream = getInputStream(ion.getContext(), request.getUri().toString());
                     if (stream == null)
                         throw new Exception("Unable to load content stream");
                     int available = stream.available();
                     InputStreamDataEmitter emitter = new InputStreamDataEmitter(ion.getHttpClient().getServer(), stream);
                     ret.setComplete(emitter);
-                    callback.onCompleted(null, new LoaderEmitter(emitter, available, LoaderEmitter.LOADED_FROM_CACHE, null, null));
+                    callback.onCompleted(null, new LoaderEmitter(emitter, available, ResponseServedFrom.LOADED_FROM_CACHE, null, null));
                 }
                 catch (Exception e) {
                     ret.setComplete(e);
